Ensure Microsoft.PowerShell.Utility is imported before module import
Automatic module import doesn't apply to types which a module exposes.
We use the Microsoft.PowerShell.Commands.BasicHtmlWebResponseObject
type, which is exposed by the Microsoft.PowerShell.Utility module, and
so must ensure it's imported before any commands we expose are run. If
not, an exception will be thrown on any invocation due to the missing
type. This also requires adding the module to ExternalModuleDependencies
as it's an internal PowerShell module, and so not in the PS Gallery.